+++ apps/nexus/nuxt.config.ts
@@ -159,7 +159,6 @@ export default defineNuxtConfig({
       email: {
         from: process.env.AUTH_EMAIL_FROM,
         server: process.env.AUTH_EMAIL_SERVER,
-        resendApiKey: process.env.RESEND_API_KEY,
       },
     },
     turnstile: {